Re-use message writing and formatting logic
